templates: fix odd bug with enum value comparisons
It turns out that it is not a good idea to use `is` when comparing
integers which is what rpc enum values essentially are.

For some obscure reason 1 did not trigger `is` 1 on my Arch Linux with
Python 3.8.2 even though on other machines with the exact same Python
version it did.

Anyway, swapping `is` with `==` fixed it, and is also what is
recommended in:
https://stackoverflow.com/a/2239753/8548472