Log mavsdk_server subprocess output
Create a thread which iterates over the lines in a given pipe and applies logging function to them. Threads are particularly suited to this use case because this is an IO bound operation. When the process is closed, the pipe will run out of lines and the thread will terminate.