Use `!=` instead of `is not` for values.
This uses the `==`/`!=` operators instead of `is`/`is not` when checking
the return value of RPC calls. In Python, `is not` tests if the two
operands are the same object whereas `!=` test that they have the same
value. `is not` just happens to work for small integers due to an
implementation detail in CPython (e.g. see [StackOverflow][1]).

[1]: https://stackoverflow.com/a/133024/1976323